fix: clean up to adjust for updated context selection, harmonize organization
