+++ src/components/annual-report/report-data/ReportData.js
@@ -1,9 +1,9 @@
 import PropTypes from 'prop-types'
 import React from 'react'
 import { SectionThree } from '../section3/SectionThree.js'
-import { SectionFour } from './section4/SectionFour.js'
 import styles from './ReportData.module.css'
-
+import { SectionOne } from './section1/SectionOne.js'
+import { SectionFour } from './section4/SectionFour.js'
 
 const SectionLayout = ({ title, children }) => (
     <div>
@@ -24,12 +24,15 @@ export const ReportData = ({ reportParameters }) => {
 
     return (
         <div className={styles.reportContainer}>
+            <SectionLayout title="Domain 1 - Completeness of Reporting">
+                <SectionOne reportParameters={reportParameters} />
+            </SectionLayout>
             <SectionLayout title="Domain 3 - External Comparison">
-                <SectionThree reportParameters={reportParameters} />                
+                <SectionThree reportParameters={reportParameters} />
             </SectionLayout>
             <SectionLayout title="Domain 4 - Consistency of Population Data">
-                <SectionFour reportParameters={reportParameters} />                
-            </SectionLayout>            
+                <SectionFour reportParameters={reportParameters} />
+            </SectionLayout>
         </div>
     )
 }
